+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.1.7] - 2023-02-22
+
+* resource organize
+
+## [1.1.6] - 2023-02-22
+
+* add modal bottom sheet
+
 ## [1.1.5] - 2023-02-03
 
 * solve generic router problem